requirements.txt
streamlit
streamlit-webrtc
pandas
numpy
python-dotenv
requests
transformers
torch
